move Npc Integrations into compatibility/npc package
