+++ src/main/java/org/betonquest/betonquest/compatibility/npc/citizens/CitizensIntegrator.java
@@ -1,15 +1,15 @@
-package org.betonquest.betonquest.compatibility.citizens;
+package org.betonquest.betonquest.compatibility.npc.citizens;
 
 import org.betonquest.betonquest.BetonQuest;
 import org.betonquest.betonquest.api.feature.FeatureAPI;
 import org.betonquest.betonquest.api.logger.BetonQuestLoggerFactory;
 import org.betonquest.betonquest.compatibility.Compatibility;
 import org.betonquest.betonquest.compatibility.Integrator;
-import org.betonquest.betonquest.compatibility.citizens.event.move.CitizensMoveController;
-import org.betonquest.betonquest.compatibility.citizens.event.move.CitizensMoveEvent;
-import org.betonquest.betonquest.compatibility.citizens.event.move.CitizensMoveEventFactory;
-import org.betonquest.betonquest.compatibility.citizens.event.move.CitizensStopEventFactory;
-import org.betonquest.betonquest.compatibility.citizens.objective.NPCKillObjective;
+import org.betonquest.betonquest.compatibility.npc.citizens.event.move.CitizensMoveController;
+import org.betonquest.betonquest.compatibility.npc.citizens.event.move.CitizensMoveEvent;
+import org.betonquest.betonquest.compatibility.npc.citizens.event.move.CitizensMoveEventFactory;
+import org.betonquest.betonquest.compatibility.npc.citizens.event.move.CitizensStopEventFactory;
+import org.betonquest.betonquest.compatibility.npc.citizens.objective.NPCKillObjective;
 import org.betonquest.betonquest.compatibility.protocollib.hider.CitizensHider;
 import org.betonquest.betonquest.kernel.registry.feature.ConversationIORegistry;
 import org.betonquest.betonquest.kernel.registry.feature.FeatureRegistries;
