+++ src/main/java/org/betonquest/betonquest/compatibility/npc/citizens/event/move/CitizensMoveEvent.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.compatibility.citizens.event.move;
+package org.betonquest.betonquest.compatibility.npc.citizens.event.move;
 
 import net.citizensnpcs.api.npc.NPC;
 import org.betonquest.betonquest.api.feature.FeatureAPI;
