+++ src/main/java/org/betonquest/betonquest/compatibility/npc/fancynpcs/FancyAdapter.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.compatibility.fancynpcs;
+package org.betonquest.betonquest.compatibility.npc.fancynpcs;
 
 import de.oliver.fancynpcs.api.Npc;
 import org.betonquest.betonquest.BetonQuest;
