+++ src/main/java/org/betonquest/betonquest/compatibility/npc/fancynpcs/FancyHider.java
@@ -1,4 +1,4 @@
-package org.betonquest.betonquest.compatibility.fancynpcs;
+package org.betonquest.betonquest.compatibility.npc.fancynpcs;
 
 import de.oliver.fancynpcs.api.events.NpcSpawnEvent;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
