+++ src/main/java/org/betonquest/betonquest/compatibility/npc/fancynpcs/FancyIdentifier.java
@@ -1,6 +1,5 @@
-package org.betonquest.betonquest.compatibility.citizens;
+package org.betonquest.betonquest.compatibility.npc.fancynpcs;
 
-import net.citizensnpcs.api.npc.NPC;
 import org.betonquest.betonquest.api.profile.OnlineProfile;
 import org.betonquest.betonquest.api.quest.npc.Npc;
 import org.betonquest.betonquest.api.quest.npc.NpcReverseIdentifier;
@@ -13,47 +12,51 @@ import java.util.Map;
 import java.util.Set;
 
 /**
- * Allows to get NpcIds for a Citizens NPC.
+ * Allows to get NpcIds for a FancyNpcs Npc.
  */
-public class CitizensReverseIdentifier implements NpcReverseIdentifier {
+public class FancyIdentifier implements NpcReverseIdentifier {
+
     /**
-     * The Citizens {@link NpcID} prefix.
+     * The {@link NpcID} prefix.
      */
-    private static final String PREFIX = "citizens ";
+    protected final String prefix;
 
     /**
      * Maps the contents of ids to the ids having that content.
      */
-    private final Map<String, Set<NpcID>> idsByInstruction;
+    protected final Map<String, Set<NpcID>> idsByInstruction;
 
     /**
-     * The default constructor.
+     * Create a new Fancy Identifier.
+     *
+     * @param prefix the prefix of relevant Ids
      */
-    public CitizensReverseIdentifier() {
-        idsByInstruction = new HashMap<>();
+    public FancyIdentifier(final String prefix) {
+        this.prefix = prefix + " ";
+        this.idsByInstruction = new HashMap<>();
     }
 
     @Override
     public Set<NpcID> getIdsFromNpc(final Npc<?> npc, @Nullable final OnlineProfile profile) {
-        if (!(npc.getOriginal() instanceof final NPC original)) {
+        if (!(npc.getOriginal() instanceof final de.oliver.fancynpcs.api.Npc original)) {
             return Set.of();
         }
-        final Set<NpcID> valid = new HashSet<>();
-        final Set<NpcID> byID = idsByInstruction.get(PREFIX + original.getId());
+        final Set<NpcID> ids = new HashSet<>();
+        final Set<NpcID> byID = idsByInstruction.get(prefix + original.getData().getId());
         if (byID != null) {
-            valid.addAll(byID);
+            ids.addAll(byID);
         }
-        final Set<NpcID> byName = idsByInstruction.get(PREFIX + original.getName() + " byName");
+        final Set<NpcID> byName = idsByInstruction.get(prefix + npc.getName() + " byName");
         if (byName != null) {
-            valid.addAll(byName);
+            ids.addAll(byName);
         }
-        return valid;
+        return ids;
     }
 
     @Override
     public void addID(final NpcID npcId) {
         final String instruction = npcId.getInstruction().toString();
-        if (instruction.startsWith(PREFIX)) {
+        if (instruction.startsWith(prefix)) {
             idsByInstruction.computeIfAbsent(instruction, string -> new HashSet<>()).add(npcId);
         }
     }
