+++ src/main/java/org/betonquest/betonquest/compatibility/npc/fancynpcs/package-info.java
@@ -1,4 +1,4 @@
 /**
  * FancyNpcs Implementation of the {@link org.betonquest.betonquest.api.quest.npc.Npc BetonQuest Npc}.
  */
-package org.betonquest.betonquest.compatibility.fancynpcs;
+package org.betonquest.betonquest.compatibility.npc.fancynpcs;
